Trying to fix xz command for windows os
xz does not exists in windows and cannot find a way to solve it using npm libraries. The possible solution I have found is to change the spawn command in windows OS to "compact [...]". Not sure if there is a problem not to passing the same arguments as for the command "xz" for linux.